requirements.txt
pyyaml==6.0
regex==2022.7.25
jinja2==3.1.2